introduce 'foundation' package
