+++ .eslintrc.js
@@ -23,6 +23,7 @@ module.exports = {
   plugins: ['@typescript-eslint', 'import'],
   rules: {
     'no-unused-vars': 'off',
+    'no-use-before-define': 'off',
     '@typescript-eslint/no-explicit-any': 'off',
     '@typescript-eslint/no-unused-vars': ['error']
   }
