+++ packages/foundation/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "@anticrm/foundation",
+  "version": "0.2.0",
+  "main": "src/index.ts",
+  "author": "Anticrm Platform Contributors",
+  "license": "EPL-2.0",
+  "scripts": {
+    "build": "tsc",
+    "build:docs": "api-extractor run --local && copyfiles -u 1 ./_api-extractor-temp/doc-models/* ../../_api-extractor-temp/",
+    "test": "echo 'no tests'",
+    "lint": "echo not defined yet"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.12.10",
+    "@babel/preset-env": "^7.12.11",
+    "@babel/preset-typescript": "^7.12.7",
+    "@microsoft/api-extractor": "^7.13.0",
+    "@types/jest": "^26.0.20",
+    "babel-jest": "^26.6.3",
+    "copyfiles": "^2.4.1",
+    "jest": "^26.6.3",
+    "typescript": "^3.9.4"
+  }
+}
